Separate tests for `sqrt_hessian` and `sqrt_hessian_sampled`
