+++ test/test_sqrt_hessian.py
@@ -3,7 +3,7 @@
 
 import pytest
 import torch
-from torch.nn import CrossEntropyLoss  # , MSELoss
+from torch.nn import CrossEntropyLoss, MSELoss
 
 from backpack import extend
 from backpack.core.derivatives import derivatives_for
@@ -74,8 +74,17 @@ SETTINGS = [
         classification_targets(shape=(1,), num_classes=1),
         (),
     ],
-    # [MSELoss(reduction="mean"), (1, 5), None],
-    # [MSELoss(reduction="sum"), (2, 6), None],
+    # reduction mean
+    # [MSELoss(reduction="mean"), (5, 1), regression_targets(shape=(5, 1)), ()],
+    # [MSELoss(reduction="mean"), (5, 2), regression_targets(shape=(5, 2)), ()],
+    # [MSELoss(reduction="sum"), (5, 1), regression_targets(shape=(5, 1)), ()],
+    # non-scalar outputs are not supported
+    [
+        MSELoss(reduction="none"),
+        (5, 1),
+        regression_targets(shape=(5, 1)),
+        (ValueError,),
+    ],
 ]
 IDS = [make_id(layer, input_shape) for (layer, input_shape, _, _) in SETTINGS]
 
@@ -166,30 +175,6 @@ def backpack_hessian_via_sqrt_hessian(layer, input, targets):
     return hessian
 
 
-def backpack_hessian_via_sqrt_hessian_sampled(layer, input, targets):
-    MC_SAMPLES = 100000
-
-    layer = extend(layer)
-    derivative = derivative_from_layer(layer)
-
-    # forward pass to initialize backpack buffers
-    _ = layer(input, targets)
-
-    sqrt_hessian = derivative.sqrt_hessian_sampled(
-        layer, None, None, mc_samples=MC_SAMPLES
-    )
-    individual_hessians = sample_hessians_from_sqrt(sqrt_hessian)
-
-    # embed
-    # TODO improve readability
-    hessian_shape = (*input.shape, *input.shape)
-    hessian = torch.zeros(hessian_shape)
-
-    hessian = embed(individual_hessians, input)
-
-    return hessian
-
-
 def sample_hessians_from_sqrt(sqrt):
     """Convert individual matrix square root into individual full matrix. """
     equation = None
@@ -229,30 +214,5 @@ def _compare_hessian_via_sqrt_hessian(layer, input, targets):
 
     check_sizes(autograd_result, backpack_result)
     check_values(autograd_result, backpack_result)
-
-    return backpack_result
-
-
-@pytest.mark.parametrize(ARGS, SETTINGS, ids=IDS)
-def test_sqrt_hessian_via_input_hessian_sampled(layer, input_shape, targets, raises):
-    """Compare the Hessian to reconstruction from individual sampled Hessian sqrt."""
-    torch.manual_seed(0)
-    input = generate_data_input_hessian(input_shape)
-    try:
-        return _compare_input_hessian_sampled(layer, input, targets)
-    except Exception as e:
-        if isinstance(e, raises):
-            return
-        else:
-            raise e
-
-
-def _compare_input_hessian_sampled(layer, input, targets):
-    RTOL = 1e-2
-    backpack_result = backpack_hessian_via_sqrt_hessian_sampled(layer, input, targets)
-    autograd_result = autograd_input_hessian(layer, input, targets)
-
-    check_sizes(autograd_result, backpack_result)
-    check_values(autograd_result, backpack_result, rtol=RTOL)
 
     return backpack_result
